refactor model structure
