+++ app/src/main/java/com/example/furstychristmas/domain/info/model/InfoPageContent.kt
@@ -1,3 +1,3 @@
-package com.example.furstychristmas.info.domain.model
+package com.example.furstychristmas.domain.info.model
 
 data class InfoPageContent(val title: String, val imageid: Int?, val text: String)
