+++ app/src/main/java/com/example/furstychristmas/domain/workout/model/Drill.kt
@@ -1,4 +1,4 @@
-package com.example.furstychristmas.workout.domain.model
+package com.example.furstychristmas.domain.workout.model
 
 import com.example.furstychristmas.model.Execution
 import com.example.furstychristmas.model.Exercise
