+++ app/src/main/java/com/example/furstychristmas/domain/workout/usecase/LoadWorkoutUseCase.kt
@@ -1,6 +1,6 @@
-package com.example.furstychristmas.workout.domain.usecase
+package com.example.furstychristmas.domain.workout.usecase
 
-import com.example.furstychristmas.workout.domain.repository.WorkoutRepository
+import com.example.furstychristmas.domain.workout.repository.WorkoutRepository
 import java.time.LocalDate
 
 class LoadWorkoutUseCase(private val workoutRepository: WorkoutRepository) {
