+++ app/src/main/java/com/example/furstychristmas/koin/ChristmasModule.kt
@@ -2,17 +2,17 @@ package com.example.furstychristmas.koin
 
 import android.content.Context
 import androidx.room.Room
-import com.example.furstychristmas.day.domain.repository.DayCompletionRepository
-import com.example.furstychristmas.day.domain.usecase.AddDayCompletionUseCase
-import com.example.furstychristmas.day.domain.usecase.DayCompletionStatusUseCase
-import com.example.furstychristmas.info.domain.repository.InfoRepository
-import com.example.furstychristmas.info.domain.usecase.LoadInfoUseCase
+import com.example.furstychristmas.domain.day.repository.DayCompletionRepository
+import com.example.furstychristmas.domain.day.usecase.AddDayCompletionUseCase
+import com.example.furstychristmas.domain.day.usecase.DayCompletionStatusUseCase
+import com.example.furstychristmas.domain.info.repository.InfoRepository
+import com.example.furstychristmas.domain.info.usecase.LoadInfoUseCase
+import com.example.furstychristmas.domain.workout.repository.WorkoutRepository
+import com.example.furstychristmas.domain.workout.usecase.LoadWorkoutUseCase
 import com.example.furstychristmas.persistence.DayDatabase
 import com.example.furstychristmas.screen.day.info.InfoViewModel
 import com.example.furstychristmas.screen.day.workout.WorkoutViewModel
 import com.example.furstychristmas.screen.overview.CardViewModel
-import com.example.furstychristmas.workout.domain.repository.WorkoutRepository
-import com.example.furstychristmas.workout.domain.usecase.LoadWorkoutUseCase
 import org.koin.android.ext.koin.androidApplication
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
