+++ app/src/main/java/com/example/furstychristmas/persistence/DayDatabase.kt
@@ -3,7 +3,7 @@ package com.example.furstychristmas.persistence
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.example.furstychristmas.day.domain.model.DayCompleted
+import com.example.furstychristmas.domain.day.model.DayCompleted
 
 @Database(entities = [DayCompleted::class], exportSchema = false, version = 2)
 @TypeConverters(Converter::class)
