+++ app/src/main/java/com/example/furstychristmas/screen/overview/DayOverviewAdapter.kt
@@ -8,7 +8,7 @@ import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
 import com.example.furstychristmas.R
 import com.example.furstychristmas.databinding.ChristmasCardBinding
-import com.example.furstychristmas.day.domain.model.DayCompleted
+import com.example.furstychristmas.domain.day.model.DayCompleted
 import com.example.furstychristmas.util.DateUtil
 import com.example.furstychristmas.util.NavigationHelper
