+++ app/src/main/java/com/example/furstychristmas/util/Util.kt
@@ -2,7 +2,7 @@ package com.example.furstychristmas.util
 
 import android.content.res.AssetManager
 import android.util.Log
-import com.example.furstychristmas.workout.domain.model.Drill
+import com.example.furstychristmas.domain.workout.model.Drill
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.Types
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
