+++ app/src/main/res/layout/exercise.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="drill"
-            type="com.example.furstychristmas.workout.domain.model.Drill" />
+            type="com.example.furstychristmas.domain.workout.model.Drill" />
     </data>
 
     <LinearLayout
